backend/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
cffi==1.14.6
cryptography==3.4.7
Django==3.2.6
django-colorfield==0.4.2
django-cors-headers==3.8.0
django-filter==2.4.0
django-rest-passwordreset==1.2.0
djangorestframework==3.12.4
jwt==1.2.0
Markdown==3.3.4
Pillow==8.3.1
pycodestyle==2.7.0
pycparser==2.20
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
